refactor: rename package from pure to common
