+++ src/main/kotlin/com/kroffle/knitting/common/exception/BaseException.kt
@@ -1,4 +1,4 @@
-package com.kroffle.knitting.pure.exception
+package com.kroffle.knitting.common.exception
 
 abstract class BaseException(override val message: String) : Exception(message) {
     abstract val httpStatus: HttpStatus
