+++ src/main/kotlin/com/kroffle/knitting/controller/ControllerException.kt
@@ -1,8 +1,8 @@
 package com.kroffle.knitting.controller
 
-import com.kroffle.knitting.pure.exception.BaseException
-import com.kroffle.knitting.pure.exception.ExceptionLayer
-import com.kroffle.knitting.pure.exception.HttpStatus
+import com.kroffle.knitting.common.exception.BaseException
+import com.kroffle.knitting.common.exception.ExceptionLayer
+import com.kroffle.knitting.common.exception.HttpStatus
 
 open class ControllerException(
     override val message: String,
