+++ src/main/kotlin/com/kroffle/knitting/controller/handler/helper/exception/UnknownException.kt
@@ -1,7 +1,7 @@
 package com.kroffle.knitting.controller.handler.helper.exception
 
+import com.kroffle.knitting.common.exception.HttpStatus
 import com.kroffle.knitting.controller.ControllerException
-import com.kroffle.knitting.pure.exception.HttpStatus
 
 class UnknownException : ControllerException(
     message = "Unknown exception raised",
