+++ src/main/kotlin/com/kroffle/knitting/domain/DomainException.kt
@@ -1,8 +1,8 @@
 package com.kroffle.knitting.domain
 
-import com.kroffle.knitting.pure.exception.BaseException
-import com.kroffle.knitting.pure.exception.ExceptionLayer
-import com.kroffle.knitting.pure.exception.HttpStatus
+import com.kroffle.knitting.common.exception.BaseException
+import com.kroffle.knitting.common.exception.ExceptionLayer
+import com.kroffle.knitting.common.exception.HttpStatus
 
 open class DomainException(override val message: String) : BaseException(message = message) {
     override val layer: ExceptionLayer = ExceptionLayer.DOMAIN
