+++ src/main/kotlin/com/kroffle/knitting/infra/InfraException.kt
@@ -1,7 +1,7 @@
 package com.kroffle.knitting.infra
 
-import com.kroffle.knitting.pure.exception.BaseException
-import com.kroffle.knitting.pure.exception.ExceptionLayer
+import com.kroffle.knitting.common.exception.BaseException
+import com.kroffle.knitting.common.exception.ExceptionLayer
 
 abstract class InfraException(override val message: String) : BaseException(message = message) {
     override val layer: ExceptionLayer = ExceptionLayer.INFRA
