+++ src/main/kotlin/com/kroffle/knitting/usecase/UseCaseException.kt
@@ -1,7 +1,7 @@
 package com.kroffle.knitting.usecase
 
-import com.kroffle.knitting.pure.exception.BaseException
-import com.kroffle.knitting.pure.exception.ExceptionLayer
+import com.kroffle.knitting.common.exception.BaseException
+import com.kroffle.knitting.common.exception.ExceptionLayer
 
 abstract class UseCaseException(
     override val message: String,
